Tested (on the frontend) and implemented a middleware auth guard, that requires a token
Haven't tested if it verifies the tokens properly, but regardless >> the fact that it checks for tokens allows us to continue frontend work and defer to that until we get to the backend sprint

Additionally, created very basic dashboardController for the moment to ensure everything works.

The data will be populated from the information from the backend.

I might also end-up allowing users to make posts on top of blogs >> the earlier v1 called posts as blogs which HAS to be taken into account to avoid any confusion.

The frontend WILL NOT BE COMPATIBLE WITH THE V1 EXPRESS BACKEND FROM THIS COMMIT ONWARDS