Created the private profile route (my-profile on the frontend, that fetches user's personal profile)
Removed created/updatedAt property from the response, it was for the user model >> not for the profile model.

It would've given the wrong information as it tracked when the user information was changed, vs the profile information.

Better yet, probably better to completely exclude it and keep it strictly for internal debugging